+++ yash-env/src/system/file_system.rs
@@ -168,45 +168,94 @@ pub enum FileType {
     Other,
 }
 
-/// File status
+/// Metadata of a file
 ///
-/// This type is a collection of file status information. It is similar to the
-/// `stat` structure defined in the POSIX standard, but it is simplified and
-/// does not include all fields of the `stat` structure.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[non_exhaustive]
-pub struct Stat {
+/// Implementations of this trait represent metadata of a file, such as its
+/// type, permissions, owner, size, etc. The [`Fstat`] trait provides methods to
+/// retrieve `Stat` objects for files.
+pub trait Stat {
     /// Device ID
-    pub dev: u64,
+    #[must_use]
+    fn dev(&self) -> u64;
     /// Inode number
-    pub ino: u64,
-    /// Access permissions
+    #[must_use]
+    fn ino(&self) -> u64;
+    /// File mode (permission bits)
     ///
     /// Note that this field does not include the file type bits.
-    /// The file type is stored in the `type` field.
-    pub mode: Mode,
+    /// Use [`type`](Self::type) to get the file type.
+    /// You can also use [`is_regular_file`](Self::is_regular_file) and other
+    /// similar methods to check the file type.
+    #[must_use]
+    fn mode(&self) -> Mode;
     /// File type
-    pub r#type: FileType,
+    #[must_use]
+    fn r#type(&self) -> FileType;
     /// Number of hard links
-    pub nlink: u64,
+    #[must_use]
+    fn nlink(&self) -> u64;
     /// User ID of the file owner
-    pub uid: Uid,
+    #[must_use]
+    fn uid(&self) -> Uid;
     /// Group ID of the file owner
-    pub gid: Gid,
-    /// Length of the file in bytes
-    pub size: u64,
+    #[must_use]
+    fn gid(&self) -> Gid;
+    /// Size of the file in bytes
+    #[must_use]
+    fn size(&self) -> u64;
     // TODO: atime, mtime, ctime, (birthtime)
-}
 
-impl Stat {
-    /// Returns the device ID and inode number as a tuple
+    /// Returns the device ID and inode number as a tuple.
     ///
     /// This method is useful for testing whether two `Stat` objects refer to
     /// the same file.
-    #[inline]
+    #[inline(always)]
     #[must_use]
-    pub const fn identity(&self) -> (u64, u64) {
-        (self.dev, self.ino)
+    fn identity(&self) -> (u64, u64) {
+        (self.dev(), self.ino())
+    }
+
+    /// Whether the file is a regular file
+    #[inline(always)]
+    #[must_use]
+    fn is_regular_file(&self) -> bool {
+        self.r#type() == FileType::Regular
+    }
+    /// Whether the file is a directory
+    #[inline(always)]
+    #[must_use]
+    fn is_directory(&self) -> bool {
+        self.r#type() == FileType::Directory
+    }
+    /// Whether the file is a symbolic link
+    #[inline(always)]
+    #[must_use]
+    fn is_symlink(&self) -> bool {
+        self.r#type() == FileType::Symlink
+    }
+    /// Whether the file is a pipe
+    #[inline(always)]
+    #[must_use]
+    fn is_fifo(&self) -> bool {
+        self.r#type() == FileType::Fifo
+    }
+    /// Whether the file is a block device
+    #[inline(always)]
+    #[must_use]
+    fn is_block_device(&self) -> bool {
+        self.r#type() == FileType::BlockDevice
+    }
+    /// Whether the file is a character device
+    #[inline(always)]
+    #[must_use]
+    fn is_character_device(&self) -> bool {
+        self.r#type() == FileType::CharacterDevice
+    }
+    /// Whether the file is a socket
+    #[inline(always)]
+    #[must_use]
+    fn is_socket(&self) -> bool {
+        self.r#type() == FileType::Socket
     }
 }
 
@@ -214,13 +263,16 @@ impl Stat {
 ///
 /// See also [`IsExecutableFile`].
 pub trait Fstat {
+    /// Metadata type returned by [`fstat`](Self::fstat) and [`fstatat`](Self::fstatat)
+    type Stat: Stat + Clone + Debug;
+
     /// Retrieves metadata of a file.
     ///
     /// This method wraps the [`fstat` system
     /// call](https://pubs.opengroup.org/onlinepubs/9799919799/functions/fstat.html).
     /// It takes a file descriptor and returns a `Stat` object containing the
     /// file metadata.
-    fn fstat(&self, fd: Fd) -> Result<Stat>;
+    fn fstat(&self, fd: Fd) -> Result<Self::Stat>;
 
     /// Retrieves metadata of a file.
     ///
@@ -230,20 +282,19 @@ pub trait Fstat {
     /// whether to follow symbolic links. It returns a `Stat` object containing
     /// the file metadata. The file path is interpreted relative to the
     /// directory represented by the directory file descriptor.
-    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<Stat>;
+    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<Self::Stat>;
 
     /// Whether there is a directory at the specified path.
     #[must_use]
     fn is_directory(&self, path: &CStr) -> bool {
         self.fstatat(AT_FDCWD, path, /* follow_symlinks */ true)
-            .is_ok_and(|stat| stat.r#type == FileType::Directory)
+            .is_ok_and(|stat| stat.is_directory())
     }
 
     /// Tests if a file descriptor is a pipe.
     #[must_use]
     fn fd_is_pipe(&self, fd: Fd) -> bool {
-        self.fstat(fd)
-            .is_ok_and(|stat| stat.r#type == FileType::Fifo)
+        self.fstat(fd).is_ok_and(|stat| stat.is_fifo())
     }
 }