+++ yash-env/src/system/virtual.rs
@@ -107,7 +107,6 @@ use super::Sigaction;
 use super::Sigmask;
 use super::SigmaskOp;
 use super::Signals;
-use super::Stat;
 use super::Sysconf;
 use super::TcGetPgrp;
 use super::TcSetPgrp;
@@ -361,6 +360,8 @@ impl Default for VirtualSystem {
 }
 
 impl Fstat for VirtualSystem {
+    type Stat = Stat;
+
     fn fstat(&self, fd: Fd) -> Result<Stat> {
         self.with_open_file_description(fd, |ofd| Ok(ofd.file.borrow().stat()))
     }