+++ yash-env-test-helper/CHANGELOG.md
@@ -9,12 +9,12 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
-## [0.10.1] - Unreleased
+## [0.11.0] - Unreleased
 
 ### Changed
 
 - Public dependency versions:
-    - yash-env 0.12.0 → 0.12.1
+    - yash-env 0.12.0 → 0.13.0
 
 ## [0.10.0] - 2026-02-04
 
@@ -101,7 +101,7 @@ A _private dependency_ is used internally and not visible to downstream users.
 
 - Initial implementation of the `yash-env-test-helper` crate
 
-[0.10.1]: https://github.com/magicant/yash-rs/releases/tag/yash-env-test-helper-0.10.1
+[0.11.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-test-helper-0.11.0
 [0.10.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-test-helper-0.10.0
 [0.9.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-test-helper-0.9.0
 [0.8.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-test-helper-0.8.0